% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IDR2}
\alias{IDR2}
\title{Subject identifier for the rye metabolomics regression tutorial (no individuals are repeated twice)}
\usage{
data(freelive)
}
\description{
Subject identifier for the rye metabolomics regression tutorial (no individuals are repeated twice)
}
\keyword{data}
